Refactor delegation rights handling and improve code organization
