+++ src/features/amUI/common/DelegationModal/utils/rightsUtils.ts
