+++ src/features/amUI/requestPage/DraftRequestPage/DraftRequestPageContent.tsx
@@ -17,7 +17,7 @@ import classes from './DraftRequestPage.module.css';
 import { usePartyRepresentation } from '../../common/PartyRepresentationContext/PartyRepresentationContext';
 import { ResourceHeading } from '../../common/DelegationModal/SingleRights/ResourceHeading';
 import { RightsSection } from '../../common/DelegationModal/SingleRights/RightsSection';
-import { useRightsSection } from '../../common/DelegationModal/SingleRights/hooks/useRightsSection';
+import { useSingleRightsDelegationRightsData } from '../../common/DelegationModal/SingleRights/hooks/useSingleRightsDelegationRightsData';
 import { DelegationAction } from '../../common/DelegationModal/EditModal';
 
 interface DraftRequestPageContentProps {
@@ -47,7 +47,10 @@ export const DraftRequestPageContent = ({ request }: DraftRequestPageContentProp
     { data: withdrawResponse, error: withdrawRequestError, isLoading: isWithdrawingRequest },
   ] = useWithdrawRequestMutation();
 
-  const { rights, setRights } = useRightsSection({ resource: request.resource, isRequest: true });
+  const { rights, setRights } = useSingleRightsDelegationRightsData({
+    resource: request.resource,
+    isRequest: true,
+  });
 
   const onConfirmRequest = () => {
     if (!isActionButtonDisabled) {
